+++ src/Writer/Extractor/Graph.php
@@ -1,0 +1,122 @@
+<?php
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2015 Marco Muths
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+namespace PhpDA\Writer\Extractor;
+
+use Fhaculty\Graph\Edge\Directed;
+use Fhaculty\Graph\Graph as FhacultyGraph;
+use Fhaculty\Graph\Vertex;
+
+class Graph implements ExtractionInterface
+{
+    /** @var array */
+    private $data;
+
+    public function extract(FhacultyGraph $graph)
+    {
+        $this->data = array(
+            'edges'    => array(),
+            'vertices' => array(),
+            'groups'   => $graph->getAttribute('graphviz.groups', array()),
+            'label'    => $graph->getAttribute('graphviz.graph.label'),
+        );
+
+        $edges = $graph->getEdges();
+        foreach ($edges as $edge) {
+            /** @var Directed $edge */
+            $this->addEdge($edge);
+            $this->addVertex($edge->getVertexStart());
+            $this->addVertex($edge->getVertexEnd());
+        }
+
+        return $this->data;
+    }
+
+    /**
+     * @param Directed $edge
+     */
+    private function addEdge(Directed $edge)
+    {
+        $id = $edge->getVertexStart()->getId() . '=>' . $edge->getVertexEnd()->getId();
+        $this->data['edges'][$id] = $this->extractEdge($edge);
+    }
+
+    /**
+     * @param Vertex $vertex
+     */
+    private function addVertex(Vertex $vertex)
+    {
+        $id = $vertex->getId();
+        if (!array_key_exists($id, $this->data['vertices'])) {
+            $this->data['vertices'][$id] = $this->extractVertex($vertex);
+        }
+    }
+
+    /**
+     * @param Directed $edge
+     * @return array
+     */
+    private function extractEdge(Directed $edge)
+    {
+        return array(
+            'from'                       => $edge->getVertexStart()->getId(),
+            'to'                         => $edge->getVertexEnd()->getId(),
+            'locations'                  => $this->extractLocations($edge->getAttribute('locations', array())),
+            'belongsToCycle'             => $edge->getAttribute('belongsToCycle', false),
+            'referenceValidatorMessages' => $edge->getAttribute('referenceValidatorMessages'),
+        );
+    }
+
+    /**
+     * @param Vertex $vertex
+     * @return array
+     */
+    private function extractVertex(Vertex $vertex)
+    {
+        $locations = $this->extractLocations($vertex->getAttribute('locations', array()));
+
+        return array(
+            'name'        => $vertex->getId(),
+            'usedByCount' => $vertex->getEdgesIn()->count(),
+            'adt'         => $vertex->getAttribute('adt', array()),
+            'location'    => array_shift($locations),
+            'group'       => $vertex->getGroup(),
+        );
+    }
+
+    /**
+     * @param \PhpDA\Entity\Location[] $locations
+     * @return array
+     */
+    private function extractLocations(array $locations)
+    {
+        $data = array();
+        foreach ($locations as $location) {
+            $data[] = $location->toArray();
+        }
+
+        return $data;
+    }
+}